[sweeping] refactored with base class
